Delete inheritance
