+++ src/utils/CMakeLists.txt
@@ -22,7 +22,6 @@ target_sources(mg-utils
     query_memory_tracker.cpp
     exponential_backoff.cpp
     compressor.cpp
-    logger.cpp
 
     PUBLIC
     FILE_SET HEADERS
